site: add system to proxy an in memory database
Each client operates on an in-memory database to avoid sqlite database-level locks. When a connection is created, it opens a "proxy", which essentially copies the disk database into the memory database.

There's also a mechanism for keeping replicating update in different connections. A single set of updates statements can be sent to all connections and update their in memory databases.